+++ src/main/java/codesquad/csvdriver/parser/InsertParser.java
@@ -1,4 +1,4 @@
-package codesquad.database.parser;
+package codesquad.csvdriver.parser;
 
 import java.util.Arrays;
 import java.util.List;
