+++ src/main/java/codesquad/csvdriver/parser/SelectParser.java
@@ -1,4 +1,4 @@
-package codesquad.database.parser;
+package codesquad.csvdriver.parser;
 
 import java.util.ArrayList;
 import java.util.Arrays;
